+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - table update performance
     - the name filter now delays 1 second before being applied
     - time to determine the selected packages to upgrade reduced
+    - displaying the upgrade required packages in the same dialog of the upgrade order
     
 ### Improvements
 - tray